# Make matrix2quaternion faster